+++ Tests/Integration/MSFT_xSQLServerAlwaysOnService.Integration.Tests.ps1
@@ -44,14 +44,25 @@ try
         Context ('When using configuration {0}' -f $configurationName) {
             It 'Should compile and apply the MOF without throwing' {
                 {
-                    # The variable $ConfigurationData was dot-sourced above.
-                    & $configurationName `
-                        -SqlInstallCredential $mockSqlInstallCredential `
-                        -OutputPath $TestDrive `
-                        -ConfigurationData $ConfigurationData
+                    $configurationParameters = @{
+                        SqlInstallCredential = $mockSqlInstallCredential
+                        OutputPath = $TestDrive
+                        # The variable $ConfigurationData was dot-sourced above.
+                        ConfigurationData = $ConfigurationData
+                    }
 
-                    Start-DscConfiguration -Path $TestDrive `
-                        -ComputerName localhost -Wait -Verbose -Force
+                    & $configurationName @configurationParameters
+
+                    $startDscConfigurationParameters = @{
+                        Path = $TestDrive
+                        ComputerName = 'localhost'
+                        Wait = $true
+                        Verbose = $true
+                        Force = $true
+                        ErrorAction = 'Stop'
+                    }
+
+                    Start-DscConfiguration @startDscConfigurationParameters
                 } | Should Not Throw
             }